+++ kura/org.eclipse.kura.linux.net/src/main/java/org/eclipse/kura/linux/net/util/IwlistCapabilityTool.java
@@ -5,7 +5,7 @@
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
- * 
+ *
  *******************************************************************************/
 package org.eclipse.kura.linux.net.util;
 
@@ -47,7 +47,7 @@ public final class IwlistCapabilityTool {
         command.setTimeout(60);
         command.setOutputStream(new ByteArrayOutputStream());
         CommandStatus status = executorService.execute(command);
-        int exitValue = (Integer) status.getExitStatus().getExitCode();
+        int exitValue = status.getExitStatus().getExitCode();
         if (!status.getExitStatus().isSuccessful()) {
             logger.warn("error executing command --- iwlist --- exit value = {}", exitValue);
             throw new KuraException(KuraErrorCode.OS_COMMAND_ERROR, String.join(" ", cmd), exitValue);
